refactor: 피드 관련 API 함수와 타입을 사용처 뷰 레이어로 이동
