+++ apps/web/src/views/feed-detail/model/types.ts
@@ -1,0 +1,5 @@
+import { type FeedPost } from '@/entities/feed';
+
+export type FeedDetail = FeedPost & {
+  isAuthor: boolean;
+};
